+++ NEWS.md
@@ -25,6 +25,10 @@
   `hash_raw_*()`, `hash_obj_*()` and `hash_file_*()` functions to calculate
   various hashes of raw vectors, R objects and files.
 
+* You can use the new `cli.default_num_colors` option to set the default
+  number of ANSI colors, only if ANSI support is otherwise detected.
+  See the details in the manual of `num_ansi_colors()`.
+
 # cli 3.1.1
 
 * `style_hyperlink()` gains a `params=` argument (#384).